Move math_util.h to lib/
There is nothing in this file that is particularly
about image processing, and it is useful to have
it available to other parts of the code
very explicitly.
