+++ imageprocess/blit.c
@@ -4,9 +4,9 @@
 
 #include "imageprocess/blit.h"
 #include "imageprocess/interpolate.h"
-#include "imageprocess/math_util.h"
 #include "imageprocess/pixel.h"
 #include "lib/logging.h"
+#include "lib/math_util.h"
 
 /**
  * Wipe a rectangular area of pixels with the defined color.
