+++ imageprocess/primitives.c
@@ -6,8 +6,8 @@
 
 #include <stdlib.h>
 
-#include "imageprocess/math_util.h"
 #include "imageprocess/primitives.h"
+#include "lib/math_util.h"
 
 Delta distance_between(Point a, Point b) {
   return (Delta){
