+++ lib/math_util.h
