+++ unpaper.h
@@ -15,9 +15,9 @@
 
 #include "constants.h"
 #include "imageprocess/interpolate.h"
-#include "imageprocess/math_util.h"
 #include "imageprocess/pixel.h"
 #include "lib/logging.h"
+#include "lib/math_util.h"
 
 /* --- preprocessor macros ------------------------------------------------ */
